amount v2: Migrate subscription_meters to BigInt
Since the subscription_meters table is so small, we can do this as an online alter table without having to create a separate column and backfill

+++ server/polar/models/subscription_meter.py
@@ -2,7 +2,7 @@ from decimal import Decimal
 from typing import TYPE_CHECKING
 from uuid import UUID
 
-from sqlalchemy import ForeignKey, Numeric, UniqueConstraint, Uuid
+from sqlalchemy import BigInteger, ForeignKey, Numeric, UniqueConstraint, Uuid
 from sqlalchemy.orm import Mapped, declared_attr, mapped_column, relationship
 from sqlalchemy.sql.sqltypes import Integer
 
@@ -19,7 +19,7 @@ class SubscriptionMeter(RecordModel):
 
     consumed_units: Mapped[Decimal] = mapped_column(Numeric, nullable=False, default=0)
     credited_units: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    amount: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    amount: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
 
     subscription_id: Mapped[UUID] = mapped_column(
         Uuid, ForeignKey("subscriptions.id", ondelete="cascade"), nullable=False
